requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
logging==0.4.9.6
